requirements.txt
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
defusedxml==0.6.0
Django==2.2.6
django-ckeditor==5.9.0
django-cool-pagination==0.3.0
django-js-asset==1.2.2
django-meta==1.6.0
django-summernote==0.8.11.4
django-watson==1.5.2
docutils==0.14
docxtpl==0.5.4
idna==2.7
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.0
mccabe==0.6.1
mock==3.0.5
oauthlib==3.1.0
panda3d==1.10.0.dev1676
Pillow==6.1.0
psycopg2==2.8.3
Pygments==2.2.0
PyJWT==1.7.1
pylint==2.1.1
pypiwin32==223
pyqtdeploy==2.2
python-docx==0.8.7
python-twitch-client==0.6.0
python3-openid==3.1.0
pytz==2018.5
pywin32==224
requests==2.19.1
requests-oauthlib==1.3.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.1.0
urllib3==1.23
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
wrapt==1.10.11
